refactor(orchestrator): extract task query workflows
